requirements.txt
# Python version
python_version == "3.10"
Flask==2.3.2
tensorflow==2.19.0
numpy==1.25.2
Pillow==10.0.0
